[WIP] Add end-to-end test for RNN Module
The test put encoder and decoder on a remote worker, and put the
LSTM module locally. The forward pass 1st looks up the embedding
remotely, then fetchs the emb result and runs it through the local
LSTM module, and finally sends the output to the remote decoder.
The backward pass should automatically traverse through all
involved parties. The optimizer takes a list of param RRefs, and
reaches each owner to update the params.

ghstack-source-id: e7700dcfd63256fe1d2e999771a596244d1cc33b
Pull Request resolved: https://github.com/pytorch/pytorch/pull/29543